Code updated based on review
